requirements.txt
# gRPC libraries
grpcio==1.51.1
grpcio-tools==1.51.1

# Flask and Gunicorn (if you have an HTTP component)
Flask==2.1.0
Werkzeug==2.2.2
gunicorn==20.1.0

# Google Cloud Libraries
google-cloud-storage
google-cloud-bigquery

# Redis (if applicable)
redis==4.5.5  # Adjust version as needed